+++ src/main/java/ezcli/modules/ezcli_core/Ezcli.java
@@ -4,7 +4,7 @@ import ezcli.modules.ezcli_core.interactive.Interactive;
 import ezcli.modules.ezcli_core.terminal.Terminal;
 
 /**
- * Entry point for program.
+ * Entry point for program. Takes care of all loading and initializing, so modules can get to work.
  */
 public class Ezcli {
 
@@ -12,20 +12,19 @@ public class Ezcli {
     public static String prompt = "   \b\b\b>> ";
 
     public static String currDir = System.getProperty("user.dir") + "/";
-    public static String userHomeDir = System.getProperty("user.home") + "/";
+    public final static String USER_HOME_DIR = System.getProperty("user.home") + "/";
 
-    public static boolean isWin;
-    public static boolean isUnix;
+    public static boolean IS_WIN;
+    public static boolean IS_UNIX;
 
     public static boolean testOutput = false;
 
     public static void main(String[] args) {
-
         setOS();
 
         initModules();
 
-        new Interactive().run(); // start program in interactive mode
+        new Interactive().run();
     }
 
     /**
@@ -35,16 +34,17 @@ public class Ezcli {
         String os = System.getProperty("os.name").toLowerCase();
 
         if (os.contains("windows"))
-            isWin = true;
+            IS_WIN = true;
         else if ("linux".equals(os) || os.contains("mac") || "sunos".equals(os) || "freebsd".equals(os))
-            isUnix = true;
+            IS_UNIX = true;
     }
 
     /**
      * Create modules to be used in program. Pass the string you want them to be mapped to in the Module hashmap.
+     * <p>
      * When the string (for example "t" for terminal module) is detected in the Interactive module, the associated
      * module will run (using the previous example, that would be the terminal module).
-     * <br></br><br></br>
+     * <p>
      * For this implementation to work, the init() method in the Module class must be called in the constructor
      * of each module.
      */