+++ src/main/java/ezcli/modules/ezcli_core/global_io/ArrowKeys.java
@@ -1,5 +1,8 @@
 package ezcli.modules.ezcli_core.global_io;
 
+/**
+ * Arrow key codes to convert input in to, for code simplicity.
+ */
 public enum ArrowKeys {
     UP, DOWN, LEFT, RIGHT, NONE
 }