+++ src/main/java/ezcli/modules/ezcli_core/global_io/Command.java
@@ -1,5 +1,8 @@
 package ezcli.modules.ezcli_core.global_io;
 
+/**
+ * System signals, such as SIGTERM and SIGKILL.
+ */
 public enum Command {
     CTRLC, CTRLZ, NONE
 }