+++ src/main/java/ezcli/modules/ezcli_core/global_io/Keys.java
@@ -1,5 +1,8 @@
 package ezcli.modules.ezcli_core.global_io;
 
+/**
+ * Enum for generalizing input for code simplicity.
+ */
 public enum Keys {
     TAB, BCKSP, NWLN, NONE
 }