+++ src/main/java/ezcli/modules/ezcli_core/util/FileAutocomplete.java
@@ -45,9 +45,9 @@ public class FileAutocomplete {
     /**
      * Sets all variables needed for attempting to autocomplete a file or directory name, and then runs the algorithm.
      *
-     * @param command Command with complete path to autocomplete (not necessarily the same as currText)
+     * @param command    Command with complete path to autocomplete (not necessarily the same as currText)
      * @param blockClear True if method should not delete the file names currently stored, which enables tab rotation
-     * @param lockTab Whether the tab key should be processed or not
+     * @param lockTab    Whether the tab key should be processed or not
      */
     public static void init(String[] command, boolean blockClear, boolean lockTab) {
 
@@ -102,7 +102,8 @@ public class FileAutocomplete {
 
     /**
      * Using a string of text representing what has been typed presently,
-     * displays all files that match the current input.
+     * displays all files that match the current input, if more than one options are available.
+     * Otherwise autocompletes the file name.
      */
     public static void fileAutocomplete() {
 
@@ -112,7 +113,8 @@ public class FileAutocomplete {
         // Whether a new list of files should be created or not
         newList = false;
 
-        path = getPath(); // Path to directory in which to find files to complete with
+        // Path to directory in which to find files to complete with
+        path = getPath();
 
         Path p = path.startsWith("/") ? Paths.get(path) : Paths.get(Ezcli.currDir + path);
         files = p.toFile().listFiles();
@@ -123,7 +125,8 @@ public class FileAutocomplete {
         if ((!endsWithSlash && !currText.startsWith("~")) && !command.endsWith(" "))
             currText = currText.split("/")[currText.split("/").length - 1];
 
-        /* If ends with slash directory and list not yet cleared from previous tab, clear,
+        /*
+         * If ends with slash directory and list not yet cleared from previous tab, clear,
          * block clear so tab rotation works and set modText to empty string,
          * so that all files in the directory are output
          */
@@ -161,41 +164,6 @@ public class FileAutocomplete {
     }
 
     /**
-     * Returns the path to the specified directory given a string.
-     *
-     * @return Path found
-     */
-    protected static String getPath() {
-        boolean startsWithSlash = originalCommand.startsWith("/") || currText.startsWith("/");
-        endsWithSlash = originalCommand.endsWith("/") || currText.endsWith("/");
-
-        StringBuilder path = new StringBuilder(currText.startsWith("~") ? Ezcli.userHomeDir : "");
-
-        if (!"".equals(path.toString())) {
-            currText = currText.substring(1);
-            if (currText.startsWith("/"))
-                currText = currText.substring(1);
-        }
-
-        // Split text at slashes to get path, so that relevant files can be autocompleted or displayed
-        String[] splitPath = currText.split("/");
-
-        if (splitPath.length > 0) {
-            // Re-create path to look in, starting at 0 if text does not start with "/", and at 1 if it does (prevents duplicate "/")
-            for (int i = startsWithSlash ? 1 : 0; (i < (splitPath.length - 1) && !endsWithSlash) || (i < splitPath.length && endsWithSlash); i++)
-                path.append(i > 0 ? "/" : "").append(splitPath[i]); // Don't make "username/" in to "/username/"
-
-            if (!"".equals(path.toString()) && !"~".equals(currText)) // If text passed is just "" do not add a "/"
-                path.append("/");
-        }
-
-        if (startsWithSlash && "".equals(path.toString())) // If path is just "/" return it (top level dir)
-            path = new StringBuilder("/");
-
-        return path.toString();
-    }
-
-    /**
      * Autocompletes a string currText using files and folders contained in path.
      * Not for rotating through a list of possibilities, this method should only run
      * when there is only one option to autocomplete.
@@ -224,8 +192,8 @@ public class FileAutocomplete {
 
     /**
      * First will display all files and directories matching 'currText' when tab key is pressed.
-     * <br></br><br></br>
-     *
+     * <br></br>
+     * <p>
      * After this has happened, provided no other keys have been pressed (excluding tab),
      * it will iterate through all the file and directory names that match 'currText', printing the
      * missing characters to the terminal so the user can see, and appending those autocompleted
@@ -251,11 +219,49 @@ public class FileAutocomplete {
             command = originalCommand + currFile.substring(startComplete);
             System.out.print(Ezcli.prompt + getCommand());
 
-            fileNames.add(currFile); // Add to end of list (rotate through list)
+            // Add to end of list (rotate through list)
+            fileNames.add(currFile);
         }
 
         if (fileNames.size() > 0 && newList)
             // Re-output command after clearing lines
             System.out.print("\n" + Ezcli.prompt + getCommand());
+    }
+
+    /**
+     * Returns the path to the specified directory given a string.
+     *
+     * @return Path found
+     */
+    protected static String getPath() {
+        boolean startsWithSlash = originalCommand.startsWith("/") || currText.startsWith("/");
+        endsWithSlash = originalCommand.endsWith("/") || currText.endsWith("/");
+
+        StringBuilder path = new StringBuilder(currText.startsWith("~") ? Ezcli.USER_HOME_DIR : "");
+
+        if (!"".equals(path.toString())) {
+            currText = currText.substring(1);
+            if (startsWithSlash)
+                currText = currText.substring(1);
+        }
+
+        // Get path from string (discard all text after the last '/')
+        for (int i = currText.length() - 1; i >= 0; i--) {
+            if (currText.charAt(i) == '/') {
+                path.append(currText.substring(0, i));
+                break;
+            } else if (i == 0)
+                path = new StringBuilder(Ezcli.currDir);
+        }
+
+        // If text passed is just "" do not add a "/"
+        if (!"".equals(path.toString()) && !"~".equals(currText))
+            path.append("/");
+
+        // If path is just "/" return it (top level dir)
+        if (startsWithSlash && "".equals(path.toString()))
+            path = new StringBuilder("/");
+
+        return path.toString();
     }
 }